Base64EncoderBase introduced
